+++ __tests__/index.js
@@ -1,34 +1,56 @@
-/* eslint-env node */
-/* eslint-disable */
+'use strict';
 
+/* eslint-disable ava/max-asserts */
+
+var eslint = require('eslint');
+var path = require('path');
+var fs = require('fs');
 var plugin = require('../index');
-var test = require('tape');
+var test = require('ava');
 
 function isObject(obj) {
-    return typeof obj === 'object' && obj !== null
+    return typeof obj === 'object' && obj !== null;
 }
 
 test('basic properties of config', function (t) {
-    t.ok(isObject(plugin.configs), 'configs is object');
-
-    t.end();
+    t.true(isObject(plugin.configs), 'configs is object');
 });
 
-test('eslint', function (t) {
-    var eslint = require('eslint');
+test('eslint and eslint cli present', function (t) {
+    t.true(isObject(eslint), 'eslint is present');
+    t.true(typeof eslint.CLIEngine === 'function', 'eslint cli engine is present');
+});
 
-    t.ok(eslint, 'eslint is present');
-    t.ok(eslint.CLIEngine, 'eslint cli engine is present');
+test('should all configs are present in export', function (t) {
+    var configDir = path.resolve(__dirname, '../lib/config');
+    var files = [];
 
-    t.end();
+    try {
+        /* eslint-disable no-sync */
+        files = fs.readdirSync(configDir);
+        /* eslint-enable no-sync */
+    } catch (error) {
+        throw error;
+    }
+
+    var actual = files
+        .filter(function (resource) {
+            return resource !== '.eslintrc.js' && resource !== 'rules';
+        })
+        .map(function (resource) {
+            return path.basename(resource, '.js');
+        });
+
+    var excepted = Object.keys(plugin.configs);
+
+    t.deepEqual(actual, excepted, 'all configs are present in export');
 });
 
 test('load ava plugin config in eslint to validate all rule syntax is correct', function (t) {
-    var eslint = require('eslint');
     var config = plugin.configs.ava;
     var hasAvaPlugin = config.plugins.indexOf('ava') !== -1;
 
-    t.ok(hasAvaPlugin, 'there is ava plugin');
+    t.true(hasAvaPlugin, 'there is ava plugin');
 
     var cli = new eslint.CLIEngine({
         useEslintrc: false,
@@ -37,18 +59,15 @@ test('load ava plugin config in eslint to validate all rule syntax is correct', 
 
     var execute = cli.executeOnText('(function () {\n    \'use strict\';\n    var foo = 0;\n\n    foo += 1;\n}());\n');
 
-    t.ok(execute, 'eslint execute is success');
-    t.ok(isObject(execute.results), 'execute results is object');
+    t.true(isObject(execute), 'eslint execute is success');
+    t.true(isObject(execute.results), 'execute results is object');
 
-    t.equal(execute.results.length, 1, 'eslint execute with one results');
-    t.equal(execute.errorCount, 0, 'eslint execute without errors');
-    t.equal(execute.warningCount, 0, 'eslint execute without warnings');
-
-    t.end();
+    t.is(execute.results.length, 1, 'eslint execute with one results');
+    t.is(execute.errorCount, 0, 'eslint execute without errors');
+    t.is(execute.warningCount, 0, 'eslint execute without warnings');
 });
 
 test('load core plugin config in eslint to validate all rule syntax is correct', function (t) {
-    var eslint = require('eslint');
     var config = plugin.configs.core;
     var cli = new eslint.CLIEngine({
         useEslintrc: false,
@@ -57,18 +76,35 @@ test('load core plugin config in eslint to validate all rule syntax is correct',
 
     var execute = cli.executeOnText('(function () {\n    \'use strict\';\n    var foo = 0;\n\n    foo += 1;\n}());\n');
 
-    t.ok(execute, 'eslint execute is success');
-    t.ok(isObject(execute.results), 'execute results is object');
+    t.true(isObject(execute), 'eslint execute is success');
+    t.true(isObject(execute.results), 'execute results is object');
 
-    t.equal(execute.results.length, 1, 'eslint execute with one results');
-    t.equal(execute.errorCount, 0, 'eslint execute without errors');
-    t.equal(execute.warningCount, 0, 'eslint execute without warnings');
+    t.is(execute.results.length, 1, 'eslint execute with one results');
+    t.is(execute.errorCount, 0, 'eslint execute without errors');
+    t.is(execute.warningCount, 0, 'eslint execute without warnings');
+});
 
-    t.end();
+test('load esnext plugin config in eslint to validate all rule syntax is correct', function (t) {
+    var config = plugin.configs.esnext;
+
+    config.extends = [];
+
+    var cli = new eslint.CLIEngine({
+        useEslintrc: false,
+        baseConfig: config
+    });
+
+    var execute = cli.executeOnText('let foo = 0;\n\n    foo += 1;\n');
+
+    t.true(isObject(execute), 'eslint execute is success');
+    t.true(isObject(execute.results), 'execute results is object');
+
+    t.is(execute.results.length, 1, 'eslint execute with one results');
+    t.is(execute.errorCount, 0, 'eslint execute without errors');
+    t.is(execute.warningCount, 0, 'eslint execute without warnings');
 });
 
 test('load lodash plugin config in eslint to validate all rule syntax is correct', function (t) {
-    var eslint = require('eslint');
     var cli = new eslint.CLIEngine({
         useEslintrc: false,
         baseConfig: plugin.configs.lodash
@@ -76,13 +112,47 @@ test('load lodash plugin config in eslint to validate all rule syntax is correct
 
     var execute = cli.executeOnText('(function () {\n    \'use strict\';\n    var foo = 0;\n\n    foo += 1;\n}());\n');
 
-    t.ok(execute, 'eslint execute is success');
-    t.ok(isObject(execute.results), 'execute results is object');
+    t.true(isObject(execute), 'eslint execute is success');
+    t.true(isObject(execute.results), 'execute results is object');
 
-    t.equal(execute.results.length, 1, 'eslint execute with one results');
-    t.equal(execute.errorCount, 0, 'eslint execute without errors');
-    t.equal(execute.warningCount, 0, 'eslint execute without warnings');
-
-    t.end();
+    t.is(execute.results.length, 1, 'eslint execute with one results');
+    t.is(execute.errorCount, 0, 'eslint execute without errors');
+    t.is(execute.warningCount, 0, 'eslint execute without warnings');
 });
-/* eslint-enable */
+
+test('load node plugin config in eslint to validate all rule syntax is correct', function (t) {
+    var cli = new eslint.CLIEngine({
+        useEslintrc: false,
+        baseConfig: plugin.configs.node
+    });
+
+    var execute = cli.executeOnText('module.export = 1;\n');
+
+    t.true(isObject(execute), 'eslint execute is success');
+    t.true(isObject(execute.results), 'execute results is object');
+
+    t.is(execute.results.length, 1, 'eslint execute with one results');
+    t.is(execute.errorCount, 0, 'eslint execute without errors');
+    t.is(execute.warningCount, 0, 'eslint execute without warnings');
+});
+
+test('load react plugin config in eslint to validate all rule syntax is correct', function (t) {
+    var config = plugin.configs.react;
+
+    config.extends = [];
+
+    var cli = new eslint.CLIEngine({
+        useEslintrc: false,
+        baseConfig: plugin.configs.react
+    });
+
+    var execute = cli.executeOnText('var React = require(\'react\');\n'
+        + 'ReactDOM.render(<h1>{\'Hello, world!\'}</h1>,document.getElementById(\'example\'));');
+
+    t.true(isObject(execute), 'eslint execute is success');
+    t.true(isObject(execute.results), 'execute results is object');
+
+    t.is(execute.results.length, 1, 'eslint execute with one results');
+    t.is(execute.errorCount, 0, 'eslint execute without errors');
+    t.is(execute.warningCount, 0, 'eslint execute without warnings');
+});